feat(service): :art: add new currencyExchange service

+++ src/Repositories/Traits/PricesTrait.php
@@ -4,6 +4,8 @@ namespace Unusualify\Modularity\Repositories\Traits;
 
 use Illuminate\Support\Arr;
 use Illuminate\Support\Facades\Request;
+use Oobook\Priceable\Models\Currency;
+use Unusualify\Modularity\Facades\CurrencyExchange;
 
 trait PricesTrait
 {
@@ -47,6 +49,9 @@ trait PricesTrait
             return;
         }
 
+        $onlyBaseCurrency = unusualConfig('services.currency_exchange.active');
+        $baseCurrency = unusualConfig('services.currency_exchange.base_currency');
+
         foreach ($this->getColumns(__TRAIT__) as $name) {
             if (isset($fields[$name])) {
                 $existingPrices = $object->prices()->where('role', $name)->get();
@@ -55,8 +60,26 @@ trait PricesTrait
                     $priceModel = isset($priceData['id'])
                         ? $existingPrices->where('id', $priceData['id'])->first()
                         : null;
-
+                    // dd($priceData, $priceModel, $this->defaultPriceData);
                     $data = array_merge_recursive_preserve($this->defaultPriceData, $priceData + ['role' => $name]);
+
+                    if($onlyBaseCurrency){
+                        foreach ([1,2,3] as $key => $id) {
+                            $_currency = Currency::find($id);
+                            if($_currency->iso_4217 !== $baseCurrency){
+                                $_data = array_merge($data, [
+                                    'display_price' => round(CurrencyExchange::convertTo($data['display_price'], $_currency->iso_4217), 2),
+                                    'currency_id' => $_currency->id
+                                ]);
+
+                                if($existingPrices->where('currency_id', $_currency->id)->count() == 0){
+                                    $object->prices()->create(Arr::except($_data, ['id']));
+                                }else{
+                                    $existingPrices->where('currency_id', $_currency->id)->first()->update(Arr::except($_data, ['id']));
+                                }
+                            }
+                        }
+                    }
 
                     if ($priceModel) {
                         // Update existing price
@@ -65,9 +88,11 @@ trait PricesTrait
                         // Create a new price
                         $object->prices()->create(Arr::except($data, ['id']));
                     }
+
+
                 }
 
-                if ($existingPrices) {
+                if ($existingPrices && ! $onlyBaseCurrency) {
                     $pricesToDelete = $existingPrices->whereNotIn('id', Arr::pluck($fields[$name], 'id'));
                     $pricesToDelete->each->delete();
                 }
@@ -84,11 +109,23 @@ trait PricesTrait
     public function getFormFieldsPricesTrait($object, $fields)
     {
         if ($object->has('prices')) {
-            $pricesByRole = $object->prices->groupBy('role');
+            $onlyBaseCurrency = unusualConfig('services.currency_exchange.active');
+
+            $query = $object->prices();
+
+            if($onlyBaseCurrency){
+                $query = $query->where('currency_id', Request::getUserCurrency()->id);
+            }
+
+            $prices = $query->get();
+            $pricesByRole = $prices->groupBy('role');
 
             foreach ($this->getColumns(__TRAIT__) as $role) {
                 if (isset($pricesByRole[$role])) {
                     $fields[$role] = $pricesByRole[$role]->map(function ($price) {
+                        // dd();
+                        // $convertedPrice = CurrencyExchange::convertTo($price->display_price, 'TRY');
+
                         return Arr::mapWithKeys(Arr::only($price->toArray(), $this->formatableColumns), function ($val, $key) {
                             if (preg_match('/display_price|price_excluding|price_including/', $key)) {
                                 return [$key => (float) $val / 100];
@@ -99,7 +136,10 @@ trait PricesTrait
                     });
                 } else {
                     $fields[$role] = [
-                        array_merge_recursive_preserve($this->defaultPriceData, ['display_price' => 0.00, 'currency_id' => Request::getUserCurrency()->id]),
+                        array_merge_recursive_preserve($this->defaultPriceData, [
+                            'display_price' => 0.00,
+                            'currency_id' => Request::getUserCurrency()->id]
+                        ),
                     ];
                 }
             }
